requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
transformers 
sentence-transformers
sentencepiece
torch